refactor: drop cli/tui package re-export shim (#2908)
