+++ src/klangk/klangk/cli/main.py
@@ -208,7 +208,7 @@ def maybe_launch_tui(ctx: typer.Context) -> None:
     if not is_interactive():
         typer.echo(ctx.get_help())
         raise typer.Exit(code=0)
-    from .tui import run_tui  # allow-deferred-import (textual, ~440ms)
+    from .tui.app import run_tui  # allow-deferred-import (textual, ~440ms)
 
     try:
         run_tui(server_url=context.server_override)
